+++ src/plugins/hr.plugin.ts
@@ -1,5 +1,5 @@
 import {StyloPlugin, StyloPluginCreateParagraphsParams} from '../types/plugin';
-import {createEmptyElement, createUneditableDiv} from '../utils/create-element.utils';
+import {createEmptyElement} from '../utils/create-element.utils';
 import {transformParagraph} from '../utils/paragraph.utils';
 
 export const hr: StyloPlugin = {
@@ -8,11 +8,8 @@ export const hr: StyloPlugin = {
   createParagraphs: async ({container, paragraph}: StyloPluginCreateParagraphsParams) => {
     const hr: HTMLHRElement = document.createElement('hr');
 
-    const element: HTMLDivElement = createUneditableDiv();
-    element.append(hr);
-
     transformParagraph({
-      elements: [element, createEmptyElement({nodeName: 'div'})],
+      elements: [hr, createEmptyElement({nodeName: 'div'})],
       paragraph,
       container,
       focus: 'last'